requirements.txt
moviepy
gunicorn
SpeechRecognition
soundfile
google-cloud-speech
google-cloud-aiplatform>=1.38
ipython
youtube-transcript-api
uvicorn==0.20.0
fastapi==0.89.1
flask
requests-toolbelt
requests==2.28.2
pathlib
python-multipart
pexpect
langchain
transformers
deeplake
pandas
gradio
PyJWT
itsdangerous
google-cloud-translate


# certifi==2023.11.17
# click==8.1.7
# decorator==4.4.2
# enum34==1.1.10
# flask==3.0.0
# Flask-PyMongo==2.3.0
# # gce-v2v @ https://storage.googleapis.com/velostrata-release/gce-v2v/gce-v2v.tar.gz
# google-api-core==2.15.0
# google-api-python-client==2.111.0
# google-auth==2.25.2
# google-auth-httplib2==0.2.0
# google-auth-oauthlib==1.2.0
# google-cloud-aiplatform==1.38.1
# google-cloud-appengine-logging==1.4.0
# google-cloud-audit-log==0.2.5
# google-cloud-bigquery==3.14.1
# google-cloud-core==2.4.1
# google-cloud-datastore==2.19.0
# google-cloud-language==2.12.0
# google-cloud-logging==3.9.0
# google-cloud-resource-manager==1.10.4
# google-cloud-spanner==3.40.1
# google-cloud-speech==2.23.0
# google-cloud-storage==2.14.0
# google-cloud-translate==3.13.0
# google-cloud-videointelligence==2.12.0
# google-cloud-vision==3.5.0
# google-crc32c==1.5.0
# google-pasta==0.2.0
# google-resumable-media==2.7.0
# googleapis-common-protos==1.62.0
# httplib2==0.22.0
# ipython==8.18.1
# ipywidgets==8.1.1
# isort==5.13.2
# itsdangerous==2.1.2
# jsonpatch==1.33
# jsonpointer==2.4
# jsonschema==4.20.0
# jsonschema-specifications==2023.12.1
# jupyterlab-widgets==3.0.9
# langchain==0.0.344
# langchain-core==0.0.8
# langsmith==0.0.67
# lazy-object-proxy==1.6.0
# Markdown==3.5.1
# MarkupSafe==2.1.3
# matplotlib==3.8.2
# matplotlib-inline==0.1.6
# moviepy==1.0.3
# multidict==6.0.4
# mypy==1.8.0
# mypy-extensions==1.0.0
# numpy==1.26.3
# oauth2client==4.1.3
# oauthlib==3.2.2
# openpyxl==3.1.2
# opt-einsum==3.3.0
# packaging==23.2
# pandas==2.1.3
# parso==0.8.3
# pexpect==4.9.0
# Pillow==10.1.0
# platformdirs==4.1.0
# pluggy==1.3.0
# proglog==0.1.10
# prompt-toolkit==3.0.43
# proto-plus==1.23.0
# protobuf==4.23.4
# psycopg2==2.9.9
# ptyprocess==0.7.0
# pure-eval==0.2.2
# py4j==0.10.9.7
# pyasn1==0.5.1
# pyasn1-modules==0.3.0
# pycodestyle==2.6.0
# pycparser==2.21
# pycurl==7.43.0.6
# pydantic==2.5.2
# pydantic-core==2.14.5
# pydocstyle==6.3.0
# pyflakes==2.2.0
# pygments==2.17.2
# PyGObject==3.38.0
# pylint==3.0.3
# pyls-mypy==0.1.8
# pymongo==4.6.1
# pyparsing==3.1.1
# pyspark==3.5.0
# python-apt==2.2.1
# python-dateutil==2.8.2
# python-jsonrpc-server==0.4.0
# python-language-server==0.35.1
# pytoolconfig==1.2.6
# pytz==2023.3.post1
# PyYAML==6.0.1
# referencing==0.32.0
# requests==2.31.0
# requests-oauthlib==1.3.1
# rope==1.11.0
# rpds-py==0.16.2
# rsa==4.9
# seaborn==0.13.0
# SecretStorage==3.3.3
# shapely==2.0.2
# six==1.16.0
# sniffio==1.3.0
# snowballstemmer==2.2.0
# soundfile==0.12.1
# SpeechRecognition==3.10.1
# SQLAlchemy==2.0.23
# sqlparse==0.4.4
# stack-data==0.6.3
# supervisor==4.2.2
# tenacity==8.2.3
# tensorboard==2.15.1
# tensorboard-data-server==0.7.2
# tensorflow==2.15.0.post1
# tensorflow-estimator==2.15.0
# tensorflow-io-gcs-filesystem==0.35.0
# termcolor==2.4.0
# toml==0.10.2
# tomli==2.0.1
# tomlkit==0.12.3
# tqdm==4.66.1
# traitlets==5.14.1
# typing-extensions==4.9.0
# typing-inspect==0.9.0
# tzdata==2023.3
# ujson==5.9.0
# uritemplate==4.1.1
# urllib3==2.1.0
# vertexai==0.0.1
# virtualenv==20.25.0
# wcwidth==0.2.12
# webencodings==0.5.1
# werkzeug==3.0.1
# widgetsnbextension==4.0.9
# wrapt==1.14.1
# yapf==0.40.2
# yarl==1.9.3
# youtube-transcript-api==0.6.2
# zipp==3.17.0
